+++ src/CodeGenerator.cpp
@@ -1688,23 +1688,7 @@ void CodeGenerator::recordRegisterWrites( const Token& token, int issueCycle )
 unsigned int CodeGenerator::ignoredImplicitWawResourcesForRemaining( std::list<Token>::const_iterator begin,
                                                                      std::list<Token>::const_iterator end ) const
 {
-	bool readsMac = false;
-	bool readsClip = false;
-	for( std::list<Token>::const_iterator i = begin; i != end; ++i )
-	{
-		if( !i->operand() )
-			continue;
-		const std::string& name = i->operand()->name();
-		readsMac = readsMac || isVuMacReader(name);
-		readsClip = readsClip || isVuClipReader(name);
-	}
-
-	unsigned int mask = VU_RESOURCE_NONE;
-	if( !readsMac )
-		mask |= VU_RESOURCE_MAC;
-	if( !readsClip )
-		mask |= VU_RESOURCE_CLIP;
-	return mask;
+	return vuIgnoredFlagWawResourcesForRemaining( begin, end );
 }
 
 void CodeGenerator::fillBranchDelaySlots( std::list<Token>& tokens ) const